// src/app/modules/electricity-predefined-rates/interfaces/get-electricity-predefined-rates.ts
export interface IGetElectricityPredefinedRates {
    id: number;
    name: string;
    singlePrice: boolean;
    greenPower: boolean;
    rate: string;
    type: string;
    marPotP1: number;
    marPotP2: number;
    marPotP3: number;
    marPotP4: number;
    marPotP5: number;
    marPotP6: number;
    marEneP1: number;
    marEneP2: number;
    marEneP3: number;
    marEneP4: number;
    marEneP5: number;
    marEneP6: number;
    type2td: string;
}